+++ src/app/shared/dia-backend/asset/downloading/dia-backend-downloading.service.ts
@@ -1,5 +1,6 @@
-import { HttpClient, HttpErrorResponse } from '@angular/common/http';
+import { HttpErrorResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
+import { first } from 'rxjs/operators';
 import { blobToBase64 } from '../../../../utils/encoding/encoding';
 import { OnConflictStrategy } from '../../../database/table/table';
 import { HttpErrorCode } from '../../../error/error.service';
@@ -22,8 +23,7 @@ export class DiaBackendAssetDownloadingService {
   constructor(
     private readonly assetRepository: DiaBackendAssetRepository,
     private readonly mediaStore: MediaStore,
-    private readonly proofRepository: ProofRepository,
-    private readonly httpClient: HttpClient
+    private readonly proofRepository: ProofRepository
   ) {}
 
   async storeRemoteCapture(
@@ -48,11 +48,9 @@ export class DiaBackendAssetDownloadingService {
     if (!diaBackendAsset.information.proof) {
       return;
     }
-    // This function is only called by storeRemoteCapture, which should always
-    // get diaBackendAsset from backend. That should mean we're always using
-    // up-to-date asset_file_thumbnail cdn link.
-    const thumbnailBlob = await this.httpClient
-      .get(diaBackendAsset.asset_file_thumbnail, { responseType: 'blob' })
+    const thumbnailBlob = await this.assetRepository
+      .downloadFile$({ id: diaBackendAsset.id, field: 'asset_file_thumbnail' })
+      .pipe(first())
       .toPromise();
     return this.mediaStore.storeThumbnail(
       diaBackendAsset.proof_hash,