+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.54.1 - 2022-04-07
+
+### Removed
+
+- Revert Download asset using CDN instead of [DIA backend `/download/` endpoint](https://dia-backend.numbersprotocol.io/api/v3/redoc/#operation/assets_download) #1362
+
 ## 0.54.0 - 2022-04-07
 
 ### Added
@@ -18,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Remove the Capture after Gift a Capture or Web 3 archive network action #1129
 - Fix wallets and NUM transfer page UI for screen width < 360px
+- Fix refresh not working when there's no Capture
 
 ## 0.53.0 - 2022-03-31
 
@@ -28,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Warn user that the asset will become public when sharing asset profile.
 - Support network action that doesn't take in any parameter.
 - Support optional network action parameters.
+- Go pro/multi select file upload.
 
 ## 0.52.1 - 2022-03-25